backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
numpy>=1.24.0
google-cloud-bigquery>=3.14.0
pandas-gbq>=0.19.0
python-multipart>=0.0.6
PyJWT>=2.8.0
bcrypt>=4.0.0
sqlalchemy>=2.0.0